# docs/contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sweetroll HTTP API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/systems/{systemId}/preview": {
      "post": {
        "operationId": "post_systems_preview",
        "parameters": [
          {
            "name": "systemId",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "snapshot",
                    "requestId"
                  ],
                  "properties": {
                    "snapshot": {
                      "type": "object",
                      "required": [
                        "snapshotId",
                        "systemId",
                        "sourceRevision",
                        "package",
                        "expiresAt"
                      ],
                      "properties": {
                        "snapshotId": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "systemId": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "sourceRevision": {
                          "type": "integer"
                        },
                        "package": {
                          "additionalProperties": true,
                          "type": "object",
                          "properties": {}
                        },
                        "expiresAt": {
                          "format": "date-time",
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "requestId"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "const": "unauthorized",
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "requestId"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "latestRevision": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "diagnostics": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "code",
                              "path",
                              "message"
                            ],
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "requestId"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "latestRevision": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "diagnostics": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "code",
                              "path",
                              "message"
                            ],
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "requestId"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "latestRevision": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "diagnostics": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "code",
                              "path",
                              "message"
                            ],
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}